backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.8.0
pydantic-settings==2.2.0
xrpl-py==4.4.0
sqlalchemy==2.0.36
aiohttp==3.10.0
pytest==8.2.0
pytest-asyncio==0.24.0
